refactor: clean up code, moved to separate components, etc.
