+++ apps/dashboard/src/components/ProductContainerOperationsComponent.vue
@@ -1,5 +1,3 @@
-<!-- eslint-disable @intlify/vue-i18n/no-raw-text -->
-<!--TODO remove ProductGridComponent, ProductDialog and product create-->
 <template>
   <Dialog v-model:visible="visible"
           :header="header"
@@ -12,7 +10,7 @@
   >
     <div @submit="handleSubmit" class="flex flex-column gap-3">
       <div class="flex gap-5 justify-content-end" v-if="state.addToContainer">
-        <p class="prop">Add existing product</p>
+        <p class="prop">{{$t('c_productContainerOperations.Existing Product')}}</p>
         <Dropdown
             :placeholder="$t('c_ContainerAssign.Select product')"
             :options="dropdownProducts"
@@ -35,28 +33,29 @@
 
         <div class="flex flex-column w-30rem justify-content-between pr-5 py-5 gap-3">
           <div class="flex flex-row justify-content-between align-items-center">
-            <label for="name">{{ $t('c_productInfoModal.Name') }}</label>
+            <label for="name">{{ $t('c_productContainerOperations.Name') }}</label>
             <div class="relative mb-3">
               <InputText id="name" class="w-18rem" v-model="name" type="text" v-bind="nameAttrs" :disabled="!edit"/>
               <br>
               <span class="error-text-name block w15-rem" v-if="closeTo && !errors.name" @click="selectCloseTo">
-                Name is close to: {{ closeTo.name }} (click to pick)</span>
+                {{ $t('c_productContainerOperations.Close To', {name: closeTo.name}) }}</span>
               <span class="error-text">{{ errors.name }}</span>
             </div>
           </div>
           <div class="flex flex-row justify-content-between align-items-center">
-            <label for="category">{{ $t('c_productInfoModal.Category') }}</label>
+            <label for="category">{{ $t('c_productContainerOperations.Category') }}</label>
             <div class="relative mb-3">
-              <Dropdown :placeholder="$t('c_productEditModal.Please select')" :options="categories" optionLabel="name"
-                        v-model="category" class="w-18rem" id="category" v-bind="categoryAttrs" :disabled="!edit"/>
+              <Dropdown :placeholder="$t('c_productContainerOperations.Please select')"
+                        :options="categories" optionLabel="name" v-model="category" class="w-18rem" id="category"
+                        v-bind="categoryAttrs" :disabled="!edit"/>
               <br>
               <span class="error-text">{{ errors.category }}</span>
             </div>
           </div>
           <div class="flex flex-row justify-content-between align-items-center">
-            <label for="vat">{{ $t('c_productEditModal.VAT') }}</label>
+            <label for="vat">{{ $t('c_productContainerOperations.VAT') }}</label>
             <div class="relative mb-3">
-              <Dropdown :placeholder="$t('c_productEditModal.Please select VAT')" :options="vatGroups"
+              <Dropdown :placeholder="$t('c_productContainerOperations.Please select VAT')" :options="vatGroups"
                         :optionLabel="(v: VatGroupResponse) => `${v.percentage}`" class="w-18rem" v-model="vat"
                         v-bind="vatAttrs" :disabled="!edit" @change="markVatEdited()"/>
               <br>
@@ -65,7 +64,7 @@
           </div>
           <div class="flex flex-row justify-content-between align-items-center">
             <label for="alcohol">{{
-                $t('c_productEditModal.Alcohol Percentage')
+                $t('c_productContainerOperations.Alcohol Percentage')
               }}</label>
             <div class="flex flex-column flex-end relative mb-3">
               <InputNumber id="alcohol" class="w-18rem" v-model="alcoholPercentage" v-bind="alcoholPercentageAttrs"
@@ -75,7 +74,7 @@
             </div>
           </div>
           <div class="flex flex-row justify-content-between align-items-center">
-            <label for="price">{{ $t('c_productEditModal.Price') }}</label>
+            <label for="price">{{ $t('c_productContainerOperations.Price') }}</label>
             <div class="flex flex-column flex-end relative mb-3">
               <InputNumber id="price" class="w-18rem" v-model="price" v-bind="priceAttrs" :max-fraction-digits="2"
                            :disabled="!edit"/>
@@ -95,40 +94,44 @@
 
           <!-- Row for Added on -->
           <div v-if="state.displayProduct" class="flex flex-row justify-content-between">
-            <h4 class="my-0">{{ $t("c_productInfoModal.Added on") }}</h4>
+            <h4 class="my-0">{{ $t("c_productContainerOperations.Added on") }}</h4>
             <p class="my-0">{{ formatDateTime(new Date(product.createdAt ? product.createdAt.toString() : '')) }}</p>
           </div>
 
           <!-- Row for Updated on -->
           <div v-if="state.displayProduct" class="flex flex-row justify-content-between">
-            <h4 class="my-0">{{ $t("c_productInfoModal.Updated on") }}</h4>
+            <h4 class="my-0">{{ $t("c_productContainerOperations.Updated on") }}</h4>
             <p class="my-0">{{ formatDateTime(new Date(product.updatedAt ? product.updatedAt.toString() : '')) }}</p>
           </div>
 
           <!-- Row for Added by -->
           <div v-if="state.displayProduct" class="flex flex-row justify-content-between">
-            <h4 class="my-0">Owner</h4>
+            <h4 class="my-0">{{ $t("c_productContainerOperations.Owner") }}</h4>
             <p class="my-0">{{ product.owner.firstName + ' ' + product.owner.lastName }}</p>
           </div>
         </div>
       </div>
       <div class="flex justify-content-end" v-if="state.addToContainer">
-        <Button type="submit" class="save-button" disabled v-if="inContainer">Product already in container</Button>
-        <Button type="submit" class="save-button" v-else @click="handleProductAdd()">Add to container</Button>
+        <Button type="submit" class="save-button" disabled v-if="inContainer">
+          {{ $t("c_productContainerOperations.Already In") }}</Button>
+        <Button type="submit" class="save-button" v-else @click="handleProductAdd()">
+          {{ $t("c_productContainerOperations.Add To Container") }}</Button>
       </div>
       <div class="flex justify-content-between" v-if="state.displayProduct">
-        <Button type="submit" class="save-button" @click="handleDeleteContainerProduct()">Delete from container</Button>
-        <Button type="submit" class="save-button" @click="handleSaveProduct()"> Save </Button>
+        <Button type="submit" class="save-button" @click="handleDeleteContainerProduct()">
+          {{ $t("c_productContainerOperations.Delete From Container") }}</Button>
+        <Button type="submit" class="save-button" @click="handleSaveProduct()">
+          {{ $t("c_productContainerOperations.Save") }}</Button>
       </div>
       <div class="flex justify-content-end" v-if="state.createProduct">
-        <Button type="submit" class="save-button" @click="handleProductAdd()"> Create </Button>
+        <Button type="submit" class="save-button" @click="handleProductAdd()">
+          {{ $t("c_productContainerOperations.Create") }}</Button>
       </div>
     </div>
   </Dialog>
 </template>
 
 <script setup lang="ts">
-// TODO merge the ProductDialogComponent and the ProductCreateComponent adn the ContainerAssignProductDialog into a signle component here.
 import { computed, type ComputedRef, onBeforeMount, type PropType, type Ref, ref, watch } from "vue";
 import type {
   BaseUserResponse,
@@ -143,7 +146,6 @@ import { useProductStore } from "@/stores/product.store";
 import { toTypedSchema } from "@vee-validate/yup";
 import * as yup from "yup";
 import { useForm } from "vee-validate";
-import { useI18n } from "vue-i18n";
 import { getProductImageSrc } from "@/utils/imageUtils";
 import Button from "primevue/button";
 import InputText from "primevue/inputtext";
@@ -154,8 +156,6 @@ import Fuse from "fuse.js";
 import FileUpload from "primevue/fileupload";
 import { useAuthStore } from "@sudosos/sudosos-frontend-common";
 import { formatDateTime } from "@/utils/formatterUtils";
-
-const { t } = useI18n();
 
 const dialog: Ref<null | any> = ref(null);
