+++ apps/dashboard/src/components/ProductGridComponent.vue
@@ -15,7 +15,6 @@
 <script setup lang="ts">
 import type { ContainerWithProductsResponse } from "@sudosos/sudosos-client";
 import ProductComponent from "@/components/ProductComponent.vue";
-import ContainerAssignProductDialog from "@/components/ContainerAssignProductDialog.vue";
 import { type Ref, ref, type PropType } from "vue";
 import ProductContainerOperationsComponent from "@/components/ProductContainerOperationsComponent.vue";
